+++ configs/clip/vit-base-p16_pt-64xb64_in1k-384px.py
@@ -1,0 +1,12 @@
+_base_ = [
+    '../_base_/models/vit-base-p16.py',
+    '../_base_/datasets/imagenet_bs64_clip_384.py',
+    '../_base_/schedules/imagenet_bs4096_AdamW.py',
+    '../_base_/default_runtime.py'
+]
+
+# model setting/mnt/lustre/lirongjie/tmp/clip_ckpt/trans_ckpt
+model = dict(backbone=dict(pre_norm=True, ), )
+
+# schedule setting
+optim_wrapper = dict(clip_grad=dict(max_norm=1.0))