+++ configs/repvgg/repvgg-B3g4_8xb32_in1k.py
@@ -1,0 +1,3 @@
+_base_ = './repvgg-B3_8xb32_in1k.py'
+
+model = dict(backbone=dict(arch='B3g4'))