+++ mmcls/models/backbones/__init__.py
@@ -1,5 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .alexnet import AlexNet
+from .beit import BEiT
 from .conformer import Conformer
 from .convmixer import ConvMixer
 from .convnext import ConvNeXt
@@ -11,11 +12,13 @@ from .densenet import DenseNet
 from .edgenext import EdgeNeXt
 from .efficientformer import EfficientFormer
 from .efficientnet import EfficientNet
+from .efficientnet_v2 import EfficientNetV2
 from .focalnet import FocalNet
 from .hornet import HorNet
 from .hrnet import HRNet
 from .inception_v3 import InceptionV3
 from .lenet import LeNet5
+from .mixmim import MixMIMTransformer
 from .mlp_mixer import MlpMixer
 from .mobilenet_v2 import MobileNetV2
 from .mobilenet_v3 import MobileNetV3
@@ -32,6 +35,7 @@ from .resnest import ResNeSt
 from .resnet import ResNet, ResNetV1c, ResNetV1d
 from .resnet_cifar import ResNet_CIFAR
 from .resnext import ResNeXt
+from .revvit import RevVisionTransformer
 from .seresnet import SEResNet
 from .seresnext import SEResNeXt
 from .shufflenet_v1 import ShuffleNetV1
@@ -40,6 +44,7 @@ from .swin_transformer import SwinTransformer
 from .swin_transformer_v2 import SwinTransformerV2
 from .t2t_vit import T2T_ViT
 from .timm_backbone import TIMMBackbone
+from .tinyvit import TinyViT
 from .tnt import TNT
 from .twins import PCPVT, SVT
 from .van import VAN
@@ -75,6 +80,7 @@ __all__ = [
     'PCPVT',
     'SVT',
     'EfficientNet',
+    'EfficientNetV2',
     'ConvNeXt',
     'HRNet',
     'ResNetV1c',
@@ -99,4 +105,8 @@ __all__ = [
     'MobileViT',
     'DaViT',
     'FocalNet',
+    'BEiT',
+    'RevVisionTransformer',
+    'MixMIMTransformer',
+    'TinyViT',
 ]