+++ mmcls/models/utils/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .attention import (BEiTAttention, ChannelMultiheadAttention,
+from .attention import (BEiTAttention, ChannelMultiheadAttention, LeAttention,
                         MultiheadAttention, ShiftWindowMSA, WindowMSA,
                         WindowMSAV2)
 from .batch_augments import CutMix, Mixup, RandomBatchAugment, ResizeMix
@@ -11,37 +11,19 @@ from .helpers import is_tracing, to_2tuple, to_3tuple, to_4tuple, to_ntuple
 from .inverted_residual import InvertedResidual
 from .layer_scale import LayerScale
 from .make_divisible import make_divisible
+from .norm import GRN, LayerNorm2d, build_norm_layer
 from .position_encoding import (ConditionalPositionEncoding,
                                 PositionEncodingFourier)
 from .se_layer import SELayer
 
 __all__ = [
-    'channel_shuffle',
-    'make_divisible',
-    'InvertedResidual',
-    'SELayer',
-    'to_ntuple',
-    'to_2tuple',
-    'to_3tuple',
-    'to_4tuple',
-    'PatchEmbed',
-    'PatchMerging',
-    'HybridEmbed',
-    'RandomBatchAugment',
-    'ShiftWindowMSA',
-    'is_tracing',
-    'MultiheadAttention',
-    'ConditionalPositionEncoding',
-    'resize_pos_embed',
-    'resize_relative_position_bias_table',
-    'ClsDataPreprocessor',
-    'Mixup',
-    'CutMix',
-    'ResizeMix',
-    'BEiTAttention',
-    'LayerScale',
-    'WindowMSA',
-    'WindowMSAV2',
-    'ChannelMultiheadAttention',
-    'PositionEncodingFourier',
+    'channel_shuffle', 'make_divisible', 'InvertedResidual', 'SELayer',
+    'to_ntuple', 'to_2tuple', 'to_3tuple', 'to_4tuple', 'PatchEmbed',
+    'PatchMerging', 'HybridEmbed', 'RandomBatchAugment', 'ShiftWindowMSA',
+    'is_tracing', 'MultiheadAttention', 'ConditionalPositionEncoding',
+    'resize_pos_embed', 'resize_relative_position_bias_table',
+    'ClsDataPreprocessor', 'Mixup', 'CutMix', 'ResizeMix', 'BEiTAttention',
+    'LayerScale', 'WindowMSA', 'WindowMSAV2', 'ChannelMultiheadAttention',
+    'PositionEncodingFourier', 'LeAttention', 'GRN', 'LayerNorm2d',
+    'build_norm_layer'
 ]