+++ mmcls/models/utils/attention.py
@@ -1,4 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
+import itertools
+
 import numpy as np
 import torch
 import torch.nn as nn
@@ -565,6 +567,8 @@ class BEiTAttention(BaseModule):
         embed_dims (int): Number of input channels.
         num_heads (int): Number of attention heads.
         window_size (tuple[int]): The height and width of the window.
+        use_rel_pos_bias (bool): Whether to use unique relative position bias,
+            if False, use shared relative position bias defined in backbone.
         bias (str): The option to add leanable bias for q, k, v. If bias is
             True, it will add leanable bias. If bias is 'qv_bias', it will only
             add leanable bias for q, v. If bias is False, it will not add bias
@@ -582,6 +586,7 @@ class BEiTAttention(BaseModule):
                  embed_dims,
                  num_heads,
                  window_size,
+                 use_rel_pos_bias,
                  bias='qv_bias',
                  qk_scale=None,
                  attn_drop_rate=0.,
@@ -601,6 +606,7 @@ class BEiTAttention(BaseModule):
             qkv_bias = False
 
         self.window_size = window_size
+        self.use_rel_pos_bias = use_rel_pos_bias
         self._init_rel_pos_embedding()
 
         self.qkv = nn.Linear(embed_dims, embed_dims * 3, bias=qkv_bias)
@@ -613,48 +619,56 @@ class BEiTAttention(BaseModule):
         self.v_bias = nn.Parameter(torch.zeros(self.embed_dims))
 
     def _init_rel_pos_embedding(self):
-        Wh, Ww = self.window_size
-        # cls to token & token 2 cls & cls to cls
-        self.num_relative_distance = (2 * Wh - 1) * (2 * Ww - 1) + 3
-        # relative_position_bias_table shape is (2*Wh-1 * 2*Ww-1 + 3, nH)
-        self.relative_position_bias_table = nn.Parameter(
-            torch.zeros(self.num_relative_distance, self.num_heads))
+        if self.use_rel_pos_bias:
+            Wh, Ww = self.window_size
+            # cls to token & token 2 cls & cls to cls
+            self.num_relative_distance = (2 * Wh - 1) * (2 * Ww - 1) + 3
+            # relative_position_bias_table shape is (2*Wh-1 * 2*Ww-1 + 3, nH)
+            self.relative_position_bias_table = nn.Parameter(
+                torch.zeros(self.num_relative_distance, self.num_heads))
 
-        # get pair-wise relative position index for
-        # each token inside the window
-        coords_h = torch.arange(Wh)
-        coords_w = torch.arange(Ww)
-        # coords shape is (2, Wh, Ww)
-        coords = torch.stack(torch_meshgrid([coords_h, coords_w]))
-        # coords_flatten shape is (2, Wh*Ww)
-        coords_flatten = torch.flatten(coords, 1)
-        relative_coords = (
-            coords_flatten[:, :, None] - coords_flatten[:, None, :])
-        # relative_coords shape is (Wh*Ww, Wh*Ww, 2)
-        relative_coords = relative_coords.permute(1, 2, 0).contiguous()
-        # shift to start from 0
-        relative_coords[:, :, 0] += Wh - 1
-        relative_coords[:, :, 1] += Ww - 1
-        relative_coords[:, :, 0] *= 2 * Ww - 1
-        relative_position_index = torch.zeros(
-            size=(Wh * Ww + 1, ) * 2, dtype=relative_coords.dtype)
-        # relative_position_index shape is (Wh*Ww, Wh*Ww)
-        relative_position_index[1:, 1:] = relative_coords.sum(-1)
-        relative_position_index[0, 0:] = self.num_relative_distance - 3
-        relative_position_index[0:, 0] = self.num_relative_distance - 2
-        relative_position_index[0, 0] = self.num_relative_distance - 1
+            # get pair-wise relative position index for
+            # each token inside the window
+            coords_h = torch.arange(Wh)
+            coords_w = torch.arange(Ww)
+            # coords shape is (2, Wh, Ww)
+            coords = torch.stack(torch_meshgrid([coords_h, coords_w]))
+            # coords_flatten shape is (2, Wh*Ww)
+            coords_flatten = torch.flatten(coords, 1)
+            relative_coords = (
+                coords_flatten[:, :, None] - coords_flatten[:, None, :])
+            # relative_coords shape is (Wh*Ww, Wh*Ww, 2)
+            relative_coords = relative_coords.permute(1, 2, 0).contiguous()
+            # shift to start from 0
+            relative_coords[:, :, 0] += Wh - 1
+            relative_coords[:, :, 1] += Ww - 1
+            relative_coords[:, :, 0] *= 2 * Ww - 1
+            relative_position_index = torch.zeros(
+                size=(Wh * Ww + 1, ) * 2, dtype=relative_coords.dtype)
+            # relative_position_index shape is (Wh*Ww, Wh*Ww)
+            relative_position_index[1:, 1:] = relative_coords.sum(-1)
+            relative_position_index[0, 0:] = self.num_relative_distance - 3
+            relative_position_index[0:, 0] = self.num_relative_distance - 2
+            relative_position_index[0, 0] = self.num_relative_distance - 1
 
-        self.register_buffer('relative_position_index',
-                             relative_position_index)
+            self.register_buffer('relative_position_index',
+                                 relative_position_index)
+        else:
+            self.window_size = None
+            self.relative_position_bias_table = None
+            self.relative_position_index = None
 
     def init_weights(self):
         super().init_weights()
-        trunc_normal_(self.relative_position_bias_table, std=0.02)
+        if self.use_rel_pos_bias:
+            trunc_normal_(self.relative_position_bias_table, std=0.02)
 
-    def forward(self, x):
+    def forward(self, x, rel_pos_bias=None):
         """
         Args:
             x (tensor): input features with shape of (num_windows*B, N, C).
+            rel_pos_bias (tensor): input relative position bias with shape of
+                (num_heads, N, N).
         """
         B, N, C = x.shape
 
@@ -678,6 +692,11 @@ class BEiTAttention(BaseModule):
             relative_position_bias = relative_position_bias.permute(
                 2, 0, 1).contiguous()  # nH, Wh*Ww, Wh*Ww
             attn = attn + relative_position_bias.unsqueeze(0)
+
+        if rel_pos_bias is not None:
+            # use shared relative position bias
+            attn = attn + rel_pos_bias
+
         attn = attn.softmax(dim=-1)
         attn = self.attn_drop(attn)
         x = (attn @ v).transpose(1, 2).reshape(B, N, C)
@@ -771,4 +790,92 @@ class ChannelMultiheadAttention(BaseModule):
 
         if self.v_shortcut:
             x = qkv[2].squeeze(1) + x
+        return x
+
+
+class LeAttention(BaseModule):
+    """LeViT Attention. Multi-head attention with attention bias,  which is
+    proposed in `LeViT: a Vision Transformer in ConvNet’s Clothing for Faster
+    Inference<https://arxiv.org/abs/2104.01136>`_
+
+    Args:
+        dim (int): Number of input channels.
+        num_heads (int): Number of attention heads. Default: 8.
+        key_dim (int): Dimension of key. Default: None.
+        attn_ratio (int): Ratio of attention heads. Default: 8.
+        resolution (tuple[int]): Input resolution. Default: (16, 16).
+        init_cfg (dict, optional): The Config for initialization.
+    """
+
+    def __init__(self,
+                 dim,
+                 key_dim,
+                 num_heads=8,
+                 attn_ratio=4,
+                 resolution=(14, 14),
+                 init_cfg=None):
+        super().__init__(init_cfg=init_cfg)
+        # (h, w)
+        assert isinstance(resolution, tuple) and len(resolution) == 2
+        self.num_heads = num_heads
+        self.scale = key_dim**-0.5
+        self.key_dim = key_dim
+        self.nh_kd = nh_kd = key_dim * num_heads
+        self.d = int(attn_ratio * key_dim)
+        self.dh = int(attn_ratio * key_dim) * num_heads
+        self.attn_ratio = attn_ratio
+        h = self.dh + nh_kd * 2
+
+        self.norm = nn.LayerNorm(dim)
+        self.qkv = nn.Linear(dim, h)
+        self.proj = nn.Linear(self.dh, dim)
+
+        points = list(
+            itertools.product(range(resolution[0]), range(resolution[1])))
+        N = len(points)
+        attention_offsets = {}
+        idxs = []
+        for p1 in points:
+            for p2 in points:
+                offset = (abs(p1[0] - p2[0]), abs(p1[1] - p2[1]))
+                if offset not in attention_offsets:
+                    attention_offsets[offset] = len(attention_offsets)
+                idxs.append(attention_offsets[offset])
+        self.attention_biases = torch.nn.Parameter(
+            torch.zeros(num_heads, len(attention_offsets)))
+        self.register_buffer(
+            'attention_bias_idxs',
+            torch.LongTensor(idxs).view(N, N),
+            persistent=False)
+
+    @torch.no_grad()
+    def train(self, mode=True):
+        super().train(mode)
+        if mode and hasattr(self, 'ab'):
+            del self.ab
+        else:
+            self.ab = self.attention_biases[:, self.attention_bias_idxs]
+
+    def forward(self, x):  # x (B,N,C)
+        B, N, _ = x.shape
+
+        # Normalization
+        x = self.norm(x)
+
+        qkv = self.qkv(x)
+        # (B, N, num_heads, d)
+        q, k, v = qkv.view(B, N, self.num_heads,
+                           -1).split([self.key_dim, self.key_dim, self.d],
+                                     dim=3)
+        # (B, num_heads, N, d)
+        q = q.permute(0, 2, 1, 3)
+        k = k.permute(0, 2, 1, 3)
+        v = v.permute(0, 2, 1, 3)
+
+        attn = ((q @ k.transpose(-2, -1)) * self.scale +
+                (self.attention_biases[:, self.attention_bias_idxs]
+                 if self.training else self.ab))
+        attn = attn.softmax(dim=-1)
+        x = (attn @ v).transpose(1, 2).reshape(B, N, self.dh)
+        x = self.proj(x)
         return x